backend/requirements/base.txt
django-admin~=3.2.9
djangorestframework
djangorestframework-jwt
django-cors-headers
django-pydenticon
pillow
